Added package for each type to test. New Utility obj.
